+++ trunk/src/LV2/gx_amp_stereo.lv2/gxamp_stereo.cpp
@@ -128,7 +128,7 @@ inline bool atomic_compare_and_exchange(T **p, T *oldv, T *newv)
 
 #include "cab_data_table.cc"
 
-////////////////////////////// MONO ////////////////////////////////////
+////////////////////////////// STEREO ////////////////////////////////////
 
 class GxPluginStereo
 {
@@ -315,11 +315,11 @@ void GxPluginStereo::init_dsp_stereo(uint32_t rate, uint32_t bufsize_)
   
   for(uint32_t i=0; i<AMP_COUNT; i++) {
         amplifier[i] = amp_model[i]();
-        amplifier[i]->set_samplerate(static_cast<unsigned int>(rate), amplifier[i]);
+        amplifier[i]->set_samplerate(rate, amplifier[i]);
     }
   for(uint32_t i=0; i<TS_COUNT; i++) {
         tonestack[i] = tonestack_model[i]();
-        tonestack[i]->set_samplerate(static_cast<unsigned int>(rate), tonestack[i]);
+        tonestack[i]->set_samplerate(rate, tonestack[i]);
     }
   
   if (bufsize )
